+++ src/org/proj/math/tensor/LazyTensor3D.java
@@ -14,12 +14,12 @@ public abstract class LazyTensor3D extends Tensor3D {
         this.values = new Matrix[alpha];
     }
 
-    public abstract BigDecimal compute (int i, int j, int k);
+    public abstract double compute (int i, int j, int k);
 
     public Matrix compute (int i) {
         return new LazyMatrix (beta, gamma) {
             @Override
-            public BigDecimal compute (int j, int k) {
+            public double compute (int j, int k) {
                 return LazyTensor3D.this.compute(i, j, k);
             }
         };
@@ -41,7 +41,7 @@ public abstract class LazyTensor3D extends Tensor3D {
     }
 
     @Override
-    public BigDecimal get (int x, int y, int z) {
+    public double get (int x, int y, int z) {
         return get(x).get(y, z);
     }
 
@@ -55,7 +55,7 @@ public abstract class LazyTensor3D extends Tensor3D {
         public abstract Matrix compute (int i);
 
         @Override
-        public BigDecimal compute (int x, int y, int z) {
+        public double compute (int x, int y, int z) {
             return compute(x).get(y, z);
         }
     }
